+++ src/scandeval/protocols.py
@@ -205,7 +205,7 @@ class ModelSetup(Protocol):
         """
         ...
 
-    def model_exists(self, model_id: str) -> bool | str:
+    def model_exists(self, model_id: str) -> bool | dict[str, str]:
         """Check whether a model exists.
 
         Args:
@@ -213,8 +213,8 @@ class ModelSetup(Protocol):
                 The model ID.
 
         Returns:
-            Whether the model exist, or the name of an extra that needs to be installed
-            to check if the model exists.
+            Whether the model exist, or a dictionary explaining why we cannot check
+            whether the model exists.
         """
         ...